GenAI/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
sentence-transformers
groq
python-dotenv
apscheduler